Made get_contrib() easier to use
Changed function get_contrib()'s requirement that the function parameter "data" has "against" as a column member. Now the function will automatically append the target feature ("against") if it's not already a part of the "data" dataframe. This is necessary for the sub-indexing according to column value to work (see lines 185 and 206).

Ensured that get_contrib() internally uses "against" as the target feature to work with rather than an external variable named "target".

These changes should have been in the original version of the function and not implemented as afterthoughts.